Pharos_API/requirements.txt
fastapi==0.118.0
httpx==0.28.1
pydantic==2.11.10
pydantic_settings==2.11.0
python-dotenv==1.1.1
uvicorn==0.37.0
